+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 #### Scripts
  - Moved config setting/viewing to settings command which will include all settings
+ - Added custom components to settings command
 
 ## Release 0.3.0 - Production Beta